Fix: https://github.com/OreCruncher/DynamicSurroundingsFabric/pull/17#issuecomment-1014811491; Add: fading water ripple at the end;

+++ src/main/java/org/orecruncher/dsurround/config/WaterRippleStyle.java
@@ -15,7 +15,7 @@ public enum WaterRippleStyle {
 
         @Override
         public float getU1(final int age) {
-            return (age / 2f) * this.DELTA;
+            return (int) (age / 2F) * this.DELTA;
         }
 
         @Override
